+++ include/Ark/Compiler/Welder.hpp
@@ -19,14 +19,15 @@
 #include <Ark/Compiler/Common.hpp>
 #include <Ark/Compiler/AST/Node.hpp>
 #include <Ark/Compiler/AST/Parser.hpp>
-#include <Ark/Compiler/ImportSolver.hpp>
-#include <Ark/Compiler/AST/Optimizer.hpp>
-#include <Ark/Compiler/Macros/Processor.hpp>
 #include <Ark/Compiler/Compiler.hpp>
+#include <Ark/Compiler/Pass.hpp>
 
 namespace Ark
 {
-    class ARK_API Welder final
+    /**
+     * @brief The welder joins all the compiler passes, being itself one since it can output an AST too
+     */
+    class ARK_API Welder final : public internal::Pass
     {
     public:
         Welder(unsigned debug, const std::vector<std::filesystem::path>& lib_env);
@@ -44,22 +45,24 @@ namespace Ark
         bool generateBytecode();
         bool saveBytecodeToFile(const std::string& filename);
 
-        [[nodiscard]] const internal::Node& ast() const noexcept;
+        [[nodiscard]] const internal::Node& ast() const noexcept override;
         [[nodiscard]] const bytecode_t& bytecode() const noexcept;
 
     private:
-        unsigned m_debug;  ///< The debug level
+        std::vector<std::filesystem::path> m_lib_env;
+
         std::filesystem::path m_root_file;
         std::vector<std::string> m_imports;
         bytecode_t m_bytecode;
 
+        std::vector<std::unique_ptr<internal::Pass>> m_passes;
         internal::Parser m_parser;
-        internal::ImportSolver m_importer;
-        internal::MacroProcessor m_macro_processor;
-        internal::Optimizer m_optimizer;
         Compiler m_compiler;
 
         bool computeAST(const std::string& filename, const std::string& code);
+
+        // HACK so that the parser can be a pass and use the loggers
+        void process([[maybe_unused]] const internal::Node&) override {}
     };
 }  // namespace Ark